feat: move duplications to barrel files
