+++ src/components/ProtectedRoute/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAppSelector } from '../../store/hooks';
+import { useAppSelector } from '../../store';
 import { LoadingSpinner } from '../../components';
 
 const ProtectedRoute: FC = () => {
