+++ src/store/index.ts
@@ -1,0 +1,14 @@
+export * from './store';
+export * from './hooks';
+export {
+  setItems,
+  setLoading as setMenuLoading,
+  setError,
+  setSelectedCategory,
+} from './menuSlice';
+export type { MenuItem } from './menuSlice';
+export {
+  setUser,
+  setLoading as setAuthLoading
+} from './authSlice';
+export * from './cartSlice';
